# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rule Proposal API",
    "description": "\n# Onboarding & User Stories\n\n- [External Project Onboarding](docs/user_stories/external_project_onboarding.md)\n- [Internal Developer Onboarding](docs/user_stories/internal_dev_onboarding.md)\n- [AI Agent Onboarding](docs/user_stories/ai_agent_onboarding.md)\n- [Full User Story Index](docs/user_stories/INDEX.md)\n\nSee these user stories for step-by-step onboarding, automation, and best practices for all client types.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/generate-token": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Generate Token",
        "operationId": "generate_token_admin_generate_token_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenGenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/errors/{error_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Error Log",
        "description": "Get error log by ID (token required).",
        "operationId": "get_error_log_admin_errors__error_id__get",
        "parameters": [
          {
            "name": "error_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Error Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TokenGenerateRequest": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "user"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "allowed_namespaces": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Namespaces"
          },
          "namespace_permissions": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Namespace Permissions"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "TokenGenerateRequest",
        "description": "Request model for generating an API token."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}